Numbers should be numbers (no quotes) in route request

+++ library/src/main/java/com/mapzen/valhalla/Router.kt
@@ -44,7 +44,7 @@ interface Router {
     fun setBiking(): Router
     fun setMultimodal(): Router
     fun setLocation(point: DoubleArray): Router
-    fun setLocation(point: DoubleArray, heading: Float): Router
+    fun setLocation(point: DoubleArray, heading: Double): Router
     fun setLocation(point: DoubleArray,
             name: String? = null,
             street: String? = null,
